% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_html.R
\name{book_html}
\alias{book_html}
\title{LibRArs book}
\usage{
book_html(..., number_sections = TRUE)
}
\arguments{
\item{...}{Arguments to \code{bookdown::html_chapters}.}

\item{number_sections}{See
\code{rmarkdown::\link{html_document}},
\code{tufte::\link[tufte]{tufte_html}}, or the documentation of the
\code{base_format} function.}
}
\value{
R Markdown output format to pass to
  \code{\link[bookdown::render_book]{render_book}}.
}
\description{
This format was adapted from the Springer manuscript package for Springer
monographs.
}
\examples{
\dontrun{
bookdown::render_book("MyArticle.Rmd")
}

}
