% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex}
\alias{book_tex}
\title{LibRArs book.}
\usage{
book_tex(..., keep_tex = TRUE, citation_package = "none")
}
\arguments{
\item{...}{Arguments to \code{bookdown::pdf_book}}
}
\value{
R Markdown output format to pass to \code{\link[bookdown::render_book]{render_book}}
}
\description{
This format was adapted from the Springer manuscript package for Springer
monographs.
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "springer_article", package = "rticles")
}

}
