% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_format}
\alias{book_tex_format}
\title{LibRArs book}
\usage{
book_tex_format(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}.}
}
\value{
R Markdown output format to pass to
  \code{\link[bookdown::render]{render}}.
}
\description{
This format was adapted from the Springer manuscript package for Springer
monographs.
}
\examples{
\dontrun{
rmarkdown::render("MyArticle.Rmd", book_tex_format())
}

}
