% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_post}
\alias{book_tex_post}
\title{Post processing function for book_tex template}
\usage{
book_tex_post(dir)
}
\arguments{
\item{dir}{The directory where the files are located.}
}
\description{
Post processing function for book_tex template
}
