% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_pre}
\alias{book_tex_pre}
\title{Pre processing function for book_tex template}
\usage{
book_tex_pre(dir)
}
\arguments{
\item{dir}{The directory where the files are located.}
}
\description{
Pre processing function for book_tex template
}
