% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{BOOKMARK_YML_FILENAME}
\alias{BOOKMARK_YML_FILENAME}
\title{Name of the file generated out of the index/toc}
\format{An object of class \code{character} of length 1.}
\usage{
BOOKMARK_YML_FILENAME
}
\description{
Name of the file generated out of the index/toc
}
\keyword{datasets}
