% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{compile_book}
\alias{compile_book}
\title{Compiles the book running preprocessor, launching bookdown and running postprocessor}
\usage{
compile_book(path, template_name)
}
\arguments{
\item{path}{The path to the directory where files are. This is the working directory where
bookdown will run and it expects to have all Rmd files and template support files too.}

\item{template_name}{The template to use.}
}
\description{
Compiles the book running preprocessor, launching bookdown and running postprocessor
}
