% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_tex.R
\name{contribution_tex_post}
\alias{contribution_tex_post}
\title{Post processing function for contribution_tex template}
\usage{
contribution_tex_post(dir, original_dir)
}
\arguments{
\item{dir}{The directory where the compilation files are located.}

\item{original_dir}{The directory from which source files originated.}
}
\description{
Post processing function for contribution_tex template
}
