% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_python_script.R
\name{create_python_script}
\alias{create_python_script}
\title{Create python script template}
\usage{
create_python_script(script_name)
}
\arguments{
\item{script_name}{Script name}
}
\description{
Create parametrised python script template
}
