% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_R_script.R
\name{create_R_script}
\alias{create_R_script}
\title{Create R script template}
\usage{
create_R_script(script_name)
}
\arguments{
\item{script_name}{Script name}
}
\description{
Create parametrised R script template
}
