% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{deploy_template}
\alias{deploy_template}
\title{Deploys resources of a template and the template itself into a
new directory where it will be possible to successfully compile the book.}
\usage{
deploy_template(template, package = NULL, dir_name = NULL,
  file_name = NULL)
}
\arguments{
\item{template}{The name of the template file to deploy. This is just the
template directory name or the full path to the directory.
The template directory is expected to be available inside
\code{rmarkdown/templates}.}

\item{package}{(Optional) Name of package where the template is located,
if not provided, the current package is used.}

\item{dir_name}{(Optional) The name of the destination folder.}

\item{file_name}{(Optional) The name of the core template file. File
'skeleton.Rmd' will be renamed to this, otherwise the name will remain
the original.}
}
\description{
Deploys resources of a template and the template itself into a
new directory where it will be possible to successfully compile the book.
}
\examples{
\dontrun{
librarstemplates::deploy_template(template="/opt/rmd/templates/report")

librarstemplates::deploy_template(template="report", package="mypackage")

librarstemplates::deploy_template(template="report",
  package="mypackage", dir_name="myreport", file_name="report.Rmd")
}

}
