% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{draft}
\alias{draft}
\title{LibRArs main rafting function.}
\usage{
draft(file, template, package = "librarstemplates", ...)
}
\arguments{
\item{file}{The name of the Rmd file to be created from the draft.}

\item{template}{The template to use in LibRArs.}

\item{...}{The params to pass to rmarkdown::draft}
}
\value{
The book in the final desired format.
}
\description{
Wrapper for same function in rmarkdown.
}
\examples{
\dontrun{
librarstemplates::draft("MyArticle.Rmd", "book_tex")
}

}
