% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{eng_numbered}
\alias{eng_numbered}
\title{An engine for numbered environments
TODO: The engine is activated in knitr by calling it "theorem",
      but it is not specific just to theorem, modify this to have a better name}
\usage{
eng_numbered(options)
}
\description{
An engine for numbered environments
TODO: The engine is activated in knitr by calling it "theorem",
      but it is not specific just to theorem, modify this to have a better name
}
