% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{eng_unumbered}
\alias{eng_unumbered}
\title{An engine for unumbered environments
TODO: The engine is activated in knitr by calling it "proof",
      but it is not specific just to theorem, modify this to have a better name}
\usage{
eng_unumbered(options)
}
\description{
An engine for unumbered environments
TODO: The engine is activated in knitr by calling it "proof",
      but it is not specific just to theorem, modify this to have a better name
}
