% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemiology.R
\name{epidemiology}
\alias{epidemiology}
\title{Epidemiology template}
\usage{
epidemiology(..., metadata = NULL, type = "pdf")
}
\arguments{
\item{metadata}{Extra yaml metadata to include. Defaults to NULL.}

\item{type}{Either "pdf" or "word". Defaults to "pdf".}
}
\description{
This function produces an R Markdown document corresponding to the Epidemiology instructions for authors
}
\examples{
rmarkdown::render("doc.Rmd", epidemiology())
}
\keyword{template}
