% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_pkg_function}
\alias{execute_pkg_function}
\title{Find the function in this package and executes it}
\usage{
execute_pkg_function(fun_name, ...)
}
\arguments{
\item{fun_name}{The function name to look for.}

\item{...}{The arguments to pass to the function invocation.}
}
\value{
The invocation result.
}
\description{
Find the function in this package and executes it
}
