% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{extract_tocfile_metadata}
\alias{extract_tocfile_metadata}
\title{Extracts relevant metadata from index.md toc file}
\usage{
extract_tocfile_metadata(text, pattern, group_no)
}
\arguments{
\item{text}{Normalized (all newlines as \n) text content.}

\item{pattern}{The regex to use.}

\item{group_no}{The group number for extraction.}
}
\value{
The metadata.
}
\description{
Extracts relevant metadata from index.md toc file
}
