% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{FILE_EXT}
\alias{FILE_EXT}
\title{Extension of files to process}
\format{An object of class \code{character} of length 1.}
\usage{
FILE_EXT
}
\description{
Extension of files to process
}
\keyword{datasets}
