% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{generate_config}
\alias{generate_config}
\title{Generates the content of the _bookdown.yml file}
\usage{
generate_config(input)
}
\arguments{
\item{input}{The input content of the index/toc file.}
}
\value{
The content of the _bookdown.yml file emitted as a list.
}
\description{
As per \link{https://bookdown.org/yihui/bookdown/configuration.html#configuration}.
}
