% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{generate_config_file}
\alias{generate_config_file}
\title{Generates the _bookdown.yml configuration file}
\usage{
generate_config_file(input, output, no_override = FALSE)
}
\arguments{
\item{input}{The path to the index/toc file to parse and translate.}

\item{output}{The path to the directory where to emit the yml file.}

\item{no_override}{A value indicating whether to allow overriding existing files.}
}
\value{
The path to the generated yml file.
}
\description{
Generates the _bookdown.yml file at the specified location.
}
