% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{generate_indexrmd}
\alias{generate_indexrmd}
\title{Generates the content of the index.Rmd file}
\usage{
generate_indexrmd(input)
}
\arguments{
\item{input}{The input content of the index/toc file.}
}
\value{
The content of the index.Rmd file emitted as a list.
}
\description{
Generates the content of the index.Rmd file
}
