% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{generate_rmd_yaml_header}
\alias{generate_rmd_yaml_header}
\title{Generates the YAML header given the configuration}
\usage{
generate_rmd_yaml_header(config, dir_path)
}
\arguments{
\item{config}{The configuration object.}

\item{dir_path}{The path to the directory where to place the generated Rmd file.}
}
\value{
The path to the generated file.
}
\description{
Generates the YAML header given the configuration
}
