% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{get_configuration}
\alias{get_configuration}
\title{Gets the configuration object out of the toc file}
\usage{
get_configuration(input)
}
\arguments{
\item{input}{The toc (md) file.}
}
\value{
The configuration object.
}
\description{
Includes:
- title
- rmd_files
- output_dir
- out_filename
- subtitle
- author
}
