% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{get_eq_block_begin_end_lines}
\alias{get_eq_block_begin_end_lines}
\title{Get the lines where an equation block starts or terminates}
\usage{
get_eq_block_begin_end_lines(x, begin = TRUE)
}
\description{
Get the lines where an equation block starts or terminates
}
