% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{get_format_call}
\alias{get_format_call}
\title{Returns the proper function call to the desired librars format}
\usage{
get_format_call(format, extra_params = "")
}
\arguments{
\item{format}{The name of the desired format function.}

\item{extra_params}{The parameters to pass to the format invocation.
Expecting value as a list: \code{param1=value1,papam2=value2,...}}
}
\value{
The proper call to the desired format.
}
\description{
Returns the proper function call to the desired librars format
}
