% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{get_label_matches_in_eq_block}
\alias{get_label_matches_in_eq_block}
\title{Gets the indices where labels are inside equation blocks}
\usage{
get_label_matches_in_eq_block(x, label_pattern, eq_begins, eq_ends)
}
\description{
Gets the indices where labels are inside equation blocks
}
