% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_function}
\alias{get_pkg_function}
\title{Find the function in this package}
\usage{
get_pkg_function(fun_name)
}
\arguments{
\item{fun_name}{The function name to look for.}
}
\value{
The function object if found.
}
\description{
Find the function in this package
}
