% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_resource}
\alias{get_pkg_resource}
\title{Gets resources from this package.}
\usage{
get_pkg_resource(...)
}
\arguments{
\item{...}{character vectors, specifying subdirectory and file(s)
    within some package.  The default, none, returns the
    root of the package.  Wildcards are not supported.}
}
\description{
Used internally to provide resources inside this package.
}
