% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_template_info}
\alias{get_template_info}
\title{Fetches the specified template's template.yaml and returns it parsed}
\usage{
get_template_info(template_name)
}
\arguments{
\item{template_name}{The name of the template to fetch.}
}
\description{
Fetches the specified template's template.yaml and returns it parsed
}
