% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_template_list}
\alias{get_template_list}
\title{Gets a list of all available templates}
\usage{
get_template_list(as.json = FALSE)
}
\arguments{
\item{as.json}{A value indicating whether JSON output should be emitted.}
}
\description{
Gets a list of all available templates
}
