% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{Homework template}
\usage{
homework(..., metadata = NULL)
}
\arguments{
\item{metadata}{Extra yaml metadata to include. Defaults to NULL.}
}
\description{
This function produces an R Markdown document for homeworkds
}
\examples{
rmarkdown::render("doc.Rmd", homework())
}
\keyword{template}
