% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{INDEX_FILENAME}
\alias{INDEX_FILENAME}
\title{Default name of the index/toc file}
\format{An object of class \code{character} of length 1.}
\usage{
INDEX_FILENAME
}
\description{
Default name of the index/toc file
}
\keyword{datasets}
