% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{INDEX_RMD_FILENAME}
\alias{INDEX_RMD_FILENAME}
\title{Name of the file generated out of the index/toc containing the front matter}
\format{An object of class \code{character} of length 1.}
\usage{
INDEX_RMD_FILENAME
}
\description{
Name of the file generated out of the index/toc containing the front matter
}
\keyword{datasets}
