% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{INTERMEDIATE_DIR_NAME}
\alias{INTERMEDIATE_DIR_NAME}
\title{Default name of the intermediate file directory}
\format{An object of class \code{character} of length 1.}
\usage{
INTERMEDIATE_DIR_NAME
}
\description{
Default name of the intermediate file directory
}
\keyword{datasets}
