% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{is_inside_block}
\alias{is_inside_block}
\title{Gets TRUE when the index is inside a block whose
opening and ending indices are provided}
\usage{
is_inside_block(i, begins, ends)
}
\description{
Gets TRUE when the index is inside a block whose
opening and ending indices are provided
}
