% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{launch_bookdown}
\alias{launch_bookdown}
\title{Launches bookdown to create the book}
\usage{
launch_bookdown(path, type)
}
\arguments{
\item{path}{The path to the directory where files are.}

\item{type}{The format type object.}
}
\value{
The path of the processed file.
}
\description{
Launches bookdown to create the book
}
