% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{launch_rmarkdown}
\alias{launch_rmarkdown}
\title{Launches Rmarkdown to create the book}
\usage{
launch_rmarkdown(path, template_name, clean_after = TRUE)
}
\arguments{
\item{path}{The path to the file to process. Rmarkdown will parse it.}

\item{template_name}{The template to use.}
}
\description{
Launches Rmarkdown to create the book
}
