% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{list_template_dirs}
\alias{list_template_dirs}
\title{Lists the template directories that are available for consumption.}
\usage{
list_template_dirs(pkg = "librarstemplates")
}
\description{
Lists the template directories that are available for consumption.
}
