% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authoring.R
\name{make}
\alias{make}
\title{LibRArs main authoring function.}
\usage{
make(input)
}
\arguments{
\item{input}{The file name representing the book to convert into
the desired format.}
}
\value{
The book in the final desired format.
}
\description{
Entry function to convert an Rmd LibRArs book into the desired format as
specified in the \code{output} parameter of the yaml front matter.
}
\examples{
\dontrun{
librars::make("MyArticle.Rmd")
}

}
