% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{merge_chapters}
\alias{merge_chapters}
\title{Merge all the chapters together}
\usage{
merge_chapters(files, dst_file_path, before = NULL, after = NULL,
  orig = files)
}
\description{
Merge all the chapters together
}
