% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{MERGED_FILENAME}
\alias{MERGED_FILENAME}
\title{Merged Rmd file name}
\format{An object of class \code{character} of length 1.}
\usage{
MERGED_FILENAME
}
\description{
Merged Rmd file name
}
\keyword{datasets}
