% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih.R
\name{nih}
\alias{nih}
\title{NIH grant template}
\usage{
nih(..., metadata = NULL, type = "pdf")
}
\arguments{
\item{metadata}{Extra yaml metadata to include. Defaults to NULL.}

\item{type}{Either "pdf" or "word". Defaults to "pdf".}
}
\description{
This function produces an R Markdown document corresponding approximately to what an NIH grant may look like
}
\examples{
rmarkdown::render("doc.Rmd", nih())
}
\keyword{template}
