% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\title{Notes template}
\usage{
notes(..., metadata = NULL)
}
\arguments{
\item{metadata}{Extra yaml metadata to include. Defaults to NULL.}
}
\description{
This function produces an R Markdown document for notetaking
}
\examples{
rmarkdown::render("doc.Rmd", notes())
}
\keyword{template}
