% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{output_format}
\alias{output_format}
\title{Wrappers for rmarkdown functions.}
\usage{
output_format(keep_md = FALSE, clean_supporting = TRUE,
  df_print = NULL, pre_knit = NULL, post_knit = NULL,
  pre_processor = NULL, intermediates_generator = NULL,
  post_processor = NULL, on_exit = NULL, base_format = NULL)
}
\description{
Wrappers for rmarkdown functions.
}
