% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tex.R
\name{post.copy_pdf_to_original_dir}
\alias{post.copy_pdf_to_original_dir}
\title{Post processing function for book_tex template}
\usage{
post.copy_pdf_to_original_dir(dir, original_dir)
}
\arguments{
\item{dir}{The directory where the compilation files are located.}

\item{original_dir}{The directory from which source files originated.}
}
\description{
Suited for post-processing.
}
