% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{post_process}
\alias{post_process}
\title{Executes final operations}
\usage{
post_process(original_dir, intermediate_dir, type)
}
\arguments{
\item{original_dir}{The original directory.}

\item{intermediate_dir}{The intermediate directory.}

\item{type}{The format type.}
}
\description{
Executes final operations
}
