% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tex.R
\name{pre.append_res_to_index_rmd}
\alias{pre.append_res_to_index_rmd}
\title{Appends the content of a resource file to index.Rmd}
\usage{
pre.append_res_to_index_rmd(dir, template_name, res_filename)
}
\arguments{
\item{dir}{The directory where the files are located.}

\item{template_name}{Name of the template to.}

\item{res_filename}{Name of the file (ext included) whose content
should be appended to index.}
}
\description{
Suited for pre-processing.
}
\details{
In this context, we need to apply the fix/hack as explained
in \link{https://github.com/rstudio/bookdown/issues/750}.
}
