% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{prepare_env}
\alias{prepare_env}
\title{Creates the intermediate dir and all needed to carry on}
\usage{
prepare_env(intermediate_dir, no_override = FALSE)
}
\arguments{
\item{intermediate_dir}{Path where to create the intermediate dir
(does not include the name of the intermediate dir).}

\item{no_override}{A value indicating whether to allow overriding existing files.}
}
\value{
The intermediate path.
}
\description{
Creates the intermediate dir and all needed to carry on
}
