% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_template.R
\name{prepare_template}
\alias{prepare_template}
\title{Ensures that a specific template's resources are deployed to the
specified directory}
\usage{
prepare_template(template_name, dir = ".")
}
\arguments{
\item{template_name}{Name of the template to use.}

\item{dir}{Path to the folder where to deploy the template and its
resources.}
}
\value{
The directory where the template has been deployed (invisibly).
}
\description{
Ensures that a specific template's resources are deployed to the
specified directory
}
\note{
An librars template consists of a directory that contains a
    description of the template and optional additional supporting
    files that are provided along with the skeleton (e.g. a logo graphic).

    The template directory should be located at \code{inst/rmarkdown/templates}
    inside this package.
    For example, a template named \code{quarterly_report} would need to
    provide the following files within the
    \code{librarstemplates/inst/rmarkdown/templates} directory:

    \code{quarterly_report/template.yaml} \cr
    \code{quarterly_report/skeleton/} \cr

    The \code{template.yaml} file should include a \code{name} and a
    \code{description} field.

    Additional files can be added to the \code{skeleton} directory, for example:

    \code{skeleton/logo.png} \cr

    These files will automatically be copied to the deployment directory containing.
}
\examples{
\dontrun{
librarstemplates::prepare_template("quarterly_report")

librarstemplates::prepare_template("quarterly_report", dir="/home/user/dir")
}
}
