% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_tex.R
\name{presentation_tex_post}
\alias{presentation_tex_post}
\title{Post processing function for presentation_tex template}
\usage{
presentation_tex_post(dir, original_dir)
}
\arguments{
\item{dir}{The directory where the compilation files are located.}

\item{original_dir}{The directory from which source files originated.}
}
\description{
Post processing function for presentation_tex template
}
