% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_tex.R
\name{presentation_tex_pre}
\alias{presentation_tex_pre}
\title{Pre processing function for presentation_tex template}
\usage{
presentation_tex_pre(dir)
}
\arguments{
\item{dir}{The directory where the files are located.}
}
\description{
In this context, we need to apply the fix/hack as explained
in \link{https://github.com/rstudio/bookdown/issues/750}.
}
