% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{process_rmd_file}
\alias{process_rmd_file}
\title{Parses the content of a file and generates the Rmd that will be fed to bookdown}
\usage{
process_rmd_file(input)
}
\arguments{
\item{input}{The content of the file to parse and translate.}
}
\value{
The modified content to replace.
}
\description{
Parses the content of a file and generates the Rmd that will be fed to bookdown
}
