% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{process_rmd_files}
\alias{process_rmd_files}
\title{Processes all Rmd files in a directory}
\usage{
process_rmd_files(files)
}
\arguments{
\item{files}{The vector of file paths to process.}
}
\value{
The list of the paths of processed files.
}
\description{
The function will change the original files.
}
