% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.any_char_incl_single_newline}
\alias{regex.any_char_incl_single_newline}
\title{Any character or a single newline (a newline immediately followed by a non-newline)
(2 groups)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.any_char_incl_single_newline
}
\description{
Any character or a single newline (a newline immediately followed by a non-newline)
(2 groups)
}
\keyword{datasets}
