% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.env_body}
\alias{regex.env_body}
\title{Any character or a single newline zero or more times (3 groups)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.env_body
}
\description{
Any character or a single newline zero or more times (3 groups)
}
\keyword{datasets}
