% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.env_end}
\alias{regex.env_end}
\title{Any number of newline followed by a # or two blank lines (3 newlines) (3 groups)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.env_end
}
\description{
Any number of newline followed by a # or two blank lines (3 newlines) (3 groups)
}
\keyword{datasets}
