% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.env_head_no_title}
\alias{regex.env_head_no_title}
\title{Sharp followed by bang, then optional spacing, then env name (label with no spaces),
(1 group)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.env_head_no_title
}
\description{
Sharp followed by bang, then optional spacing, then env name (label with no spaces),
(1 group)
}
\keyword{datasets}
