% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.env_head_with_title}
\alias{regex.env_head_with_title}
\title{Sharp followed by bang, then optional spacing, then env name (label with no spaces),
then mandatory spacing, then title (2 groups)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.env_head_with_title
}
\description{
Sharp followed by bang, then optional spacing, then env name (label with no spaces),
then mandatory spacing, then title (2 groups)
}
\keyword{datasets}
