% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.line_start}
\alias{regex.line_start}
\title{A newline or the start of the string (1 group)}
\format{An object of class \code{character} of length 1.}
\usage{
regex.line_start
}
\description{
A newline or the start of the string (1 group)
}
\keyword{datasets}
