% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.spacing.mandatory}
\alias{regex.spacing.mandatory}
\title{A space or a tab at leats once}
\format{An object of class \code{character} of length 1.}
\usage{
regex.spacing.mandatory
}
\description{
A space or a tab at leats once
}
\keyword{datasets}
