% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.spacing.optional}
\alias{regex.spacing.optional}
\title{A space or a tab zero or more times}
\format{An object of class \code{character} of length 1.}
\usage{
regex.spacing.optional
}
\description{
A space or a tab zero or more times
}
\keyword{datasets}
