% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\docType{data}
\name{regex.spacing}
\alias{regex.spacing}
\title{A space or a tab}
\format{An object of class \code{character} of length 1.}
\usage{
regex.spacing
}
\description{
A space or a tab
}
\keyword{datasets}
