% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Renders an artifact}
\usage{
render(input, format = NULL, output = NULL, intermediate_dir = NULL,
  clean_after = TRUE, index_filename = NULL)
}
\arguments{
\item{input}{Path to the directory containing the files to process.
The directory must contain an index.md file.}

\item{format}{The format object containing info about the type of output.}

\item{output}{The directory path where to save the generated output.
If not specirfied, it will the same as input.}

\item{intermediate_dir}{The intermediate directory path where to
place the temporary files during rendering. If not specified,
it will be the same as input. The directory must exist.}

\item{clean_after}{A value indicating whether to clean the intermediate folder.}

\item{index_filename}{The name of the index file to look for. If NULL, defaults to index.md.}
}
\description{
Renders an artifact
}
