% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{retrieve_template_functions}
\alias{retrieve_template_functions}
\title{Retrieves the format, preprocessing and postprocessing functions for a template}
\usage{
retrieve_template_functions(template_name)
}
\arguments{
\item{template_name}{The name of the template.}
}
\description{
Retrieves the format, preprocessing and postprocessing functions for a template
}
