% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\name{rmdconvert.check.get.env}
\alias{rmdconvert.check.get.env}
\title{Checks whether the provided environment is supported}
\usage{
rmdconvert.check.get.env(env_name)
}
\arguments{
\item{env_name}{The name of the environment to check}
}
\value{
The environment (lower-case), otherwise error is raised.
}
\description{
If not supported, an error is raised.
}
