% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\name{rmdconvert.scan.envs}
\alias{rmdconvert.scan.envs}
\title{Scans the source string and transforms it against all environment constructs}
\usage{
rmdconvert.scan.envs(rmd_src, debug.info = FALSE)
}
\arguments{
\item{rmd_src}{The source Rmd string.}

\item{debug.info}{When \code{TRUE} it will instruct the function to log debug info.}
}
\value{
The source string with all constructs replaced.
}
\description{
Use \code{rmdconvert.scan.envs(text, TRUE)} to debug the progression of
replacements.
Using the table you get from the previous command, use
\code{stringr::str_match(<Match text>, <pattern>)} where the pattern can be
retrieved by using the variable name associated to the 'Pattern name' column
of the table.
}
\details{
An environment is created by using the syntax: \code{#!<name> <title>}. E.g.
\dontrun{
  #!theorem Euler-Lagrange equation
  The extremant functionals are to be found as the solutions of the
  Euler-Lagrange equation.
}
Which gets translated into:
\dontrun{
  ```{theorem, name = "Euler-Lagrange equation"}
  The extremant functionals are to be found as the solutions of the
  Euler-Lagrange equation.
  ```
}
}
