% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\name{rmdconvert.scan}
\alias{rmdconvert.scan}
\title{Scans the source string and transforms it against all constructs}
\usage{
rmdconvert.scan(rmd_src, debug.info = FALSE)
}
\arguments{
\item{rmd_src}{The source Rmd string.}

\item{debug.info}{When \code{TRUE} it will instruct the function to log debug info.}
}
\value{
The source string with all constructs replaced.
}
\description{
Important: the order though which target constructs are scanned
is important.
}
