% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_convert.R
\name{rmsconvert.create.env}
\alias{rmsconvert.create.env}
\title{Creates the transformed output for an environment}
\usage{
rmsconvert.create.env(name, body, title = NA, before = NA,
  after = NA)
}
\arguments{
\item{name}{The name of the environment.}

\item{body}{The content of the environment.}

\item{title}{The title of the environment.}

\item{before}{The before content.}

\item{after}{The after content.}
}
\description{
Creates the transformed output for an environment
}
