% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{slides}
\alias{slides}
\title{Slides template}
\usage{
slides(..., metadata = NULL)
}
\arguments{
\item{metadata}{Extra yaml metadata to include. Defaults to NULL.}
}
\description{
This function produces an R Markdown document corresponding to a particular Beamer template
}
\examples{
rmarkdown::render("doc.Rmd", slides())
}
\keyword{template}
