% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_excerpt}
\alias{str_excerpt}
\title{Utility to represent long string in format: 'str1...str2'}
\usage{
str_excerpt(str, lower = 10, upper = 10)
}
\description{
Utility to represent long string in format: 'str1...str2'
}
