% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template.get_path}
\alias{template.get_path}
\title{Gets the path to the template folder of the specified template}
\usage{
template.get_path(template_name, package = "librarstemplates")
}
\description{
Gets the path to the template folder of the specified template
}
