% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template}
\alias{template}
\title{Templates constructors}
\usage{
template(x, envir = parent.frame())
}
\arguments{
\item{x}{(character | expression)}

\item{envir}{(environment)}
}
\description{
Templates constructors
}

