% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_blog_template.R
\name{use_blog_template}
\alias{use_blog_template}
\title{Create folder and files for website blog post}
\usage{
use_blog_template(date_chr = "2025-12-31", slug)
}
\arguments{
\item{date_chr}{Date in yyyy-mm-dd format.}

\item{slug}{Character string with url slug e.g. \code{"blog-post-title"}}
}
\value{
a message if template file was successfully copied over
}
\description{
Create folder and files for website blog post
}
