% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rmarkdown_template2.R
\name{use_rmarkdown_template2}
\alias{use_rmarkdown_template2}
\title{Add an RMarkdown Template}
\usage{
use_rmarkdown_template2(template_name = "Template Name",
  template_dir = tolower(asciify(template_name)),
  template_description = "A description of the template",
  template_create_dir = FALSE, default_template = "markdown-template.Rmd",
  package = "usethis")
}
\arguments{
\item{template_name}{The name as printed in the template menu.}

\item{template_dir}{Name of the directory the template will live in within inst/rmarkdown/templates.}

\item{template_description}{Sets the value of description in template.yml.}

\item{template_create_dir}{Sets the value of create_dir in template.yml.}

\item{default_template}{The default template to use as the basis for the new template}

\item{package}{The package containing the default template}
}
\description{
Copy of the usethis function of the same name with added functionality
Adds files and directories necessary to add a custom rmarkdown template to RStudio.
}
