% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_skills.R
\name{use_skills}
\alias{use_skills}
\title{Create Skills.md file}
\usage{
use_skills(type, dir = ".claude/skills")
}
\arguments{
\item{type}{Type of Skills.md file. Options include \code{"r-general"}.}

\item{dir}{Directory to put files into. Default \code{".claude/skills/"}.}
}
\value{
a message if template file was successfully copied over
}
\description{
Create Skills.md file
}
