% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_viz_template.R
\name{use_viz_template}
\alias{use_viz_template}
\title{Create folder and files for data viz}
\usage{
use_viz_template(dir = ".", filename = "script", lang = "R")
}
\arguments{
\item{dir}{Directory to put script}

\item{filename}{File path where script should go inlcuding name.}

\item{lang}{Language, either \code{"R"}, \code{"Python"}, or \code{"D3"}.}
}
\value{
a message if template file was successfully copied over
}
\description{
Create folder and files for data viz
}
