% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asCharacter}
\alias{asCharacter}
\alias{asFunction}
\title{Util functions}
\usage{
asCharacter(x, ..., parent = parent.frame())

asFunction(x, parent = parent.frame())
}
\arguments{
\item{x}{(Template)}

\item{...}{dots}

\item{parent}{(environment)}
}
\description{
Util functions
}

