% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consts.R
\docType{data}
\name{YML_HEADER_FILENAME}
\alias{YML_HEADER_FILENAME}
\title{Name of the file generated from configuration representing the header of the Rmd artifact}
\format{An object of class \code{character} of length 1.}
\usage{
YML_HEADER_FILENAME
}
\description{
Name of the file generated from configuration representing the header of the Rmd artifact
}
\keyword{datasets}
