% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_format}
\alias{book_tex_format}
\title{Format function for book_tex template}
\usage{
book_tex_format(clean_after = TRUE, ...)
}
\arguments{
\item{clean_after}{A value indicating whether to clean after or not.}

\item{...}{Arguments to \code{rmarkdown::pdf_document}.}
}
\value{
R Markdown output format to pass to
  \code{\link[librarstemplates::render]{render}}.
}
\description{
This format renders a PDF and selects the template available
in the folder at the moment of rendering.
}
\examples{
\dontrun{
librarstemplates::render("./book/", librarstemplates::book_tex_format())
}

}
