% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_format}
\alias{book_tex_format}
\title{Format function for book_tex template}
\usage{
book_tex_format(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}.}
}
\value{
R Markdown output format to pass to
  \code{\link[bookdown::render_book]{render_book}}.
}
\description{
This format renders a PDF and selects the template available
in the folder at the moment of rendering.
}
\examples{
\dontrun{
bookdown::render_book("MyArticle.Rmd", librarstemplates::book_tex_format())
}

}
