% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_tex.R
\name{book_tex_post}
\alias{book_tex_post}
\title{Post processing function for book_tex template}
\usage{
book_tex_post(dir, original_dir)
}
\arguments{
\item{dir}{The directory where the compilation files are located.}

\item{original_dir}{The directory from which source files originated.}
}
\description{
Post processing function for book_tex template
}
