% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{compile_artifacts}
\alias{compile_artifacts}
\title{Compiles the artifacts running preprocessor, launching bookdown and running postprocessor}
\usage{
compile_artifacts(path, template_name, dstpath, ...)
}
\arguments{
\item{path}{The path to the directory where files are. This is the working directory where
bookdown will run and it expects to have all Rmd files and template support files too.}

\item{template_name}{The template to use.}

\item{dstpath}{The path where to move the final artifacts once generated.}

\item{...}{Parameters to pass to \code{launch_bookdown}.}
}
\description{
Compiles the artifacts running preprocessor, launching bookdown and running postprocessor
}
