% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Locates the template and deploys the 'draft' content folder
into the specified path}
\usage{
draft(template_name, dir = ".")
}
\arguments{
\item{template_name}{Name of the template to use.}

\item{dir}{Path to the folder where to deploy the draft resources.}
}
\value{
The directory where the files have been deployed (invisibly).
}
\description{
Locates the template and deploys the 'draft' content folder
into the specified path
}
\note{
A template might not provide a draft.
}
\examples{
\dontrun{
librarstemplates::draft("quarterly_report")

librarstemplates::draft("quarterly_report", dir="/home/user/dir")
}
}
