% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{launch_bookdown}
\alias{launch_bookdown}
\title{Launches bookdown to create the book}
\usage{
launch_bookdown(path, template_name, clean_after = TRUE)
}
\arguments{
\item{path}{The path to the directory where files are. This is the working directory where
bookdown will run and it expects to have all Rmd files and template support files too.}

\item{template_name}{The template to use.}
}
\description{
Launches bookdown to create the book
}
