% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{process_rmd_files}
\alias{process_rmd_files}
\title{Parses all Rmd files into bookdown Rmd}
\usage{
process_rmd_files(path)
}
\arguments{
\item{path}{The folder where to search Rmd files.}
}
\value{
The list of the paths of processed files.
}
\description{
The function will chanbge the original files.
}
