% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Renders an artifact}
\usage{
render(input, template_name = "book_tex", output = NULL,
  intermediate_dir = NULL, clean_after = TRUE, index_filename = NULL)
}
\arguments{
\item{input}{Path to the directory containing the files to process.
The directory must contain an index.md file.}

\item{template_name}{The name of the template to use. This is the name of the template
directory in the \code{templates} folder.}

\item{output}{The directory path where to save the generated output.
If not specirfied, it will the same as input.}

\item{intermediate_dir}{The directory path where the intermediate folder will be created.
The intermediate folder is the directory where to
place the temporary files during rendering. If not specified,
it will be the same as input. The directory must exist.}

\item{clean_after}{A value indicating whether to clean the intermediate folder.}

\item{index_filename}{The name of the index file to look for. If NULL, defaults to index.md.}
}
\description{
Renders an artifact
}
