% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templateChar}
\alias{templateChar}
\alias{templateExpr}
\alias{templateFun}
\title{Templates constructors}
\usage{
templateChar(x)

templateExpr(x)

templateFun(x)
}
\arguments{
\item{x}{}
}
\description{
Templates constructors
}

