% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{tmpl}
\alias{tmpl}
\title{Templates constructors}
\usage{
tmpl(x, ..., envir = parent.frame())
}
\arguments{
\item{x}{(character | expression)}

\item{...}{(name = value | name ~ value) name-value expressions}

\item{envir}{(environment)}
}
\description{
Templates constructors
}

