% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci}
\alias{use_gitlab_ci}
\title{Create GitLab CI/CD file}
\usage{
use_gitlab_ci(type, overwrite = FALSE)
}
\arguments{
\item{type}{Type of CI/CD pipeline. Options include \code{"pages"}, \code{"quarto"}, \code{"quarto-r"}, or \code{"pkgdown"}.}

\item{overwrite}{Boolean to overwrite existing file if it exists. Default \code{FALSE}.}
}
\value{
a message if template file was successfully copied over
}
\description{
Create GitLab CI/CD file
}
