% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmplUpdate}
\alias{tmplEval}
\alias{tmplUpdate}
\alias{tmplUpdate.default}
\alias{tmplUpdate.function}
\title{Util functions}
\usage{
tmplUpdate(x, ...)

\method{tmplUpdate}{default}(x, ...)

\method{tmplUpdate}{function}(x, ...)

tmplEval(x, envir = new.env(parent = parent.frame()), ...)
}
\arguments{
\item{x}{(template)}

\item{...}{dots}

\item{envir}{(environment)}
}
\description{
Util functions
}

