% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{templateEval}
\alias{templateEval}
\alias{templateEval.default}
\alias{templateEval.function}
\alias{templateEvalHere}
\alias{templateEvalLocal}
\alias{templateSub}
\alias{templateSub.default}
\title{Util functions}
\usage{
templateEval(x, ...)

\method{templateEval}{default}(x, ...)

\method{templateEval}{function}(x, ...)

templateSub(x, ...)

\method{templateSub}{default}(x, ...)

templateEvalHere(template, envir = parent.frame(), ...)

templateEvalLocal(template, envir = parent.frame(), ...)
}
\arguments{
\item{x}{(template)}

\item{...}{dots}

\item{parent}{(environment)}
}
\description{
Util functions
}

